+++ lib/winston/common.js
@@ -9,6 +9,7 @@
 var util = require('util'),
     crypto = require('crypto'),
     cycle = require('cycle'),
+    fs = require('fs'),
     config = require('./config');
 
 //
@@ -285,4 +286,62 @@ exports.serialize = function (obj, key) {
   }
 
   return msg;
+};
+
+
+//
+// ### function tailFile (options, callback)
+// #### @options {Object} Options for tail.
+// #### @callback {function} Callback to execute on every line.
+// `tail -f` a file. Options must include file.
+//
+exports.tailFile = function tail(options, callback) {
+  var stream = fs.createReadStream(options.file, { encoding: 'utf8' }),
+      buff = '',
+      destroy,
+      row = 0;
+
+  destroy = stream.destroy.bind(stream);
+  stream.destroy = function () {};
+
+  if (options.start === -1) {
+    delete options.start;
+  }
+  
+  stream.on('data', function (data) {
+    var data = (buff + data).split(/\n+/),
+        l = data.length - 1,
+        i = 0;
+
+    for (; i < l; i++) {
+      if (options.start == null || row > options.start) {
+        callback(null, data[i]);
+      }
+      row++;
+    }
+
+    buff = data[l];
+  });
+
+  stream.on('error', function (err) {
+    callback(err);
+    destroy();
+  });
+
+  stream.on('end', function () {
+    if (buff) {
+      stream.emit('line', buff);
+      buff = '';
+    }
+
+    resume();
+  });
+
+  function resume() {
+    setTimeout(function () {
+      stream.resume();
+    }, 1000);
+  }
+
+  return destroy;
 };
