+++ lib/winston/transports/file.js
@@ -276,66 +276,6 @@ File.prototype.query = function (options, callback) {
 };
 
 //
-// ### function _tail (options, callback)
-// #### @options {Object} Options for tail.
-// #### @callback {function} Callback to execute on every line.
-// `tail -f` a file. Options must include file.
-//
-File.prototype._tail = function tail(options, callback) {
-  var stream = fs.createReadStream(options.file, { encoding: 'utf8' }),
-      buff = '',
-      destroy,
-      row = 0;
-
-  destroy = stream.destroy.bind(stream);
-  stream.destroy = function () {};
-
-  if (options.start === -1) {
-    delete options.start;
-  }
-  
-  stream.on('data', function (data) {
-    var data = (buff + data).split(/\n+/),
-        l = data.length - 1,
-        i = 0;
-
-    for (; i < l; i++) {
-      if (options.start == null || row > options.start) {
-        stream.emit('line', data[i]);
-      }
-      row++;
-    }
-
-    buff = data[l];
-  });
-
-  stream.on('line', function (data) {
-    if (callback) callback(data);
-  });
-
-  stream.on('error', function (err) {
-    destroy();
-  });
-
-  stream.on('end', function () {
-    if (buff) {
-      stream.emit('line', buff);
-      buff = '';
-    }
-
-    resume();
-  });
-
-  function resume() {
-    setTimeout(function () {
-      stream.resume();
-    }, 1000);
-  }
-
-  return destroy;
-};
-
-//
 // ### function stream (options)
 // #### @options {Object} Stream options for this instance.
 // Returns a log stream for this transport. Options object is optional.
@@ -350,7 +290,12 @@ File.prototype.stream = function (options) {
     start: options.start
   };
 
-  stream.destroy = this._tail(tail, function (line) {
+  stream.destroy = common.tailFile(tail, function (err, line) {
+    
+    if(err){
+      return stream.emit('error',err);
+    }
+    
     try {
       stream.emit('data', line);
       line = JSON.parse(line);
